Add files via upload
#Module 14 Challenge
# Bike-sharing Analysis
## Overview
The objective of this project is to convince investors that a bike-sharing program in Des Moines is a solid business proposal. To solidify the proposal, using the data from an operational and successful program is key. 
Via Tableau we’ll be showing potential investors key behaviors and data from the Citi Bike program in NYC, using public data for the entire month of Aug (2019)
## Results
As you can see on the story linked below, some of the conclusions reached are:
•	Largest majority of users are male
•	Busiest days are Saturdays and Sundays
•	During week days peak hours are between 7am to 9am and at the end of the day 5pm to 7pm
•	Subscribers are the majority of the user base
•	Customers (non-subscribers) tend to have longer trips
•	Largest majority of the trips are under 40 or so mintes with a high spike around 5 minutes

•	Tableau Story can be seen here ![Story]( https://public.tableau.com/profile/ana7380#!/vizhome/Citibike_M14/CitiBikeStory?publish=yes) (Deliverable 3)
•	Tableau file (Deliverable 2) includes all requested graphs. Summary seen on this Dashboard Image ![CitiBike Dash](link)
•	Data format was modified using this Python code (Deliverable 1)  ![jupyter notebook](link)